+++ test/document.test.js
@@ -211,7 +211,7 @@ describe('document', function() {
       doc.nested.setAge = 10;
       assert.equal(doc.nested.age, 10);
       doc.nested.setr = 'set it';
-      assert.equal(doc.getValue('nested.setr'), 'set it setter');
+      assert.equal(doc.$__getValue('nested.setr'), 'set it setter');
 
       const doc2 = new TestDocument();
       doc2.init({
@@ -4183,9 +4183,10 @@ describe('document', function() {
       });
 
       schema.post('save', function(error, res, next) {
-        if (error instanceof MongooseError.DocumentNotFoundError) {
-          error = new Error('Somebody else updated the document!');
-        }
+        assert.ok(error instanceof MongooseError.DocumentNotFoundError);
+        assert.ok(error.message.indexOf('gh4004') !== -1, error.message);
+
+        error = new Error('Somebody else updated the document!');
         next(error);
       });
 
@@ -7575,6 +7576,154 @@ describe('document', function() {
     return Model.create({ name: 'foo' }).then(() => assert.ok(false), err => {
       assert.equal(err.errors['name'].message, 'Oops!');
       assert.ok(err.message.indexOf('Oops!') !== -1, err.message);
+    });
+  });
+
+  it('handles nested properties named `schema` (gh-7831)', function() {
+    const schema = new mongoose.Schema({ nested: { schema: String } });
+    const Model = db.model('gh7831', schema);
+
+    return co(function*() {
+      yield Model.collection.insertOne({ nested: { schema: 'test' } });
+
+      const doc = yield Model.findOne();
+      assert.strictEqual(doc.nested.schema, 'test');
+    });
+  });
+
+  describe('overwrite() (gh-7830)', function() {
+    let Model;
+
+    before(function() {
+      const schema = new Schema({
+        _id: Number,
+        name: String,
+        nested: {
+          prop: String
+        },
+        arr: [Number],
+        immutable: {
+          type: String,
+          immutable: true
+        }
+      });
+      Model = db.model('gh7830', schema);
+    });
+
+    it('works', function() {
+      return co(function*() {
+        const doc = yield Model.create({
+          _id: 1,
+          name: 'test',
+          nested: { prop: 'foo' },
+          immutable: 'bar'
+        });
+        doc.overwrite({ name: 'test2' });
+
+        assert.deepEqual(doc.toObject(), {
+          _id: 1,
+          __v: 0,
+          name: 'test2',
+          immutable: 'bar'
+        });
+      });
+    });
+
+    it('skips version key', function() {
+      return co(function*() {
+        yield Model.collection.insertOne({
+          _id: 2,
+          __v: 5,
+          name: 'test',
+          nested: { prop: 'foo' },
+          immutable: 'bar'
+        });
+        const doc = yield Model.findOne({ _id: 2 });
+        doc.overwrite({ _id: 2, name: 'test2' });
+
+        assert.deepEqual(doc.toObject(), {
+          _id: 2,
+          __v: 5,
+          name: 'test2',
+          immutable: 'bar'
+        });
+      });
+    });
+  });
+
+  describe('immutable properties (gh-7671)', function() {
+    let Model;
+
+    before(function() {
+      const schema = new Schema({
+        createdAt: {
+          type: Date,
+          immutable: true,
+          default: new Date('6/1/2019')
+        },
+        name: String
+      });
+      Model = db.model('gh7671', schema);
+    });
+
+    it('SchemaType#immutable()', function() {
+      const schema = new Schema({
+        createdAt: {
+          type: Date,
+          default: new Date('6/1/2019')
+        },
+        name: String
+      });
+
+      assert.ok(!schema.path('createdAt').$immutable);
+
+      schema.path('createdAt').immutable(true);
+      assert.ok(schema.path('createdAt').$immutable);
+      assert.equal(schema.path('createdAt').setters.length, 1);
+
+      schema.path('createdAt').immutable(false);
+      assert.ok(!schema.path('createdAt').$immutable);
+      assert.equal(schema.path('createdAt').setters.length, 0);
+    });
+
+    it('with save()', function() {
+      let doc = new Model({ name: 'Foo' });
+      return co(function*() {
+        assert.equal(doc.createdAt.toLocaleDateString('en-us'), '6/1/2019');
+        yield doc.save();
+
+        doc = yield Model.findOne({ createdAt: new Date('6/1/2019') });
+        doc.createdAt = new Date('6/1/2017');
+        assert.equal(doc.createdAt.toLocaleDateString('en-us'), '6/1/2019');
+
+        doc.set({ createdAt: new Date('6/1/2021') });
+        assert.equal(doc.createdAt.toLocaleDateString('en-us'), '6/1/2019');
+
+        yield doc.save();
+
+        doc = yield Model.findOne({ createdAt: new Date('6/1/2019') });
+        assert.ok(doc);
+      });
+    });
+
+    it('with update', function() {
+      let doc = new Model({ name: 'Foo' });
+      return co(function*() {
+        assert.equal(doc.createdAt.toLocaleDateString('en-us'), '6/1/2019');
+        yield doc.save();
+
+        const update = { createdAt: new Date('6/1/2020') };
+
+        yield Model.updateOne({}, update);
+
+        doc = yield Model.findOne();
+        assert.equal(doc.createdAt.toLocaleDateString('en-us'), '6/1/2019');
+
+        const err = yield Model.updateOne({}, update, { strict: 'throw' }).
+          then(() => null, err => err);
+        assert.equal(err.name, 'StrictModeError');
+        assert.ok(err.message.indexOf('createdAt') !== -1, err.message);
+      });
     });
   });
 });