+++ test/model.findOneAndUpdate.test.js
@@ -838,6 +838,24 @@ describe('model: findOneAndUpdate:', function() {
       });
     });
   });
+  it('return hydrated document (gh-7734 gh-7735)', function() {
+    const fruitSchema = new Schema({
+      name: { type: String }
+    });
+
+    const Fruit = db.model('gh-7734', fruitSchema);
+    return co(function*() {
+      let fruit = yield Fruit.create({ name: 'Apple' });
+
+      fruit = yield Fruit.findOneAndUpdate({}, { $set: { name: 'Banana' }},
+        { new: true, useFindAndModify: false });
+      assert.ok(fruit instanceof mongoose.Document);
+
+      fruit = yield Fruit.findOneAndUpdate({}, { $set: { name: 'Cherry' }},
+        { new: true, useFindAndModify: true });
+      assert.ok(fruit instanceof mongoose.Document);
+    });
+  });
   it('return rawResult when doing an upsert & new=false gh-7770', function(done) {
     const thingSchema = new Schema({
       _id: String,
@@ -2312,6 +2330,21 @@ describe('model: findOneAndUpdate:', function() {
       // Should not throw
       doc = yield Model.findOneAndUpdate({ _id: doc._id }, void 0, { new: true });
       assert.equal(doc.name, 'test');
+    });
+  });
+
+  it('runs lowercase on $addToSet, $push, etc (gh-4185)', function() {
+    const Cat = db.model('gh4185', {
+      _id: String,
+      myArr: { type: [{type: String, lowercase: true}], default: undefined }
+    });
+
+    return co(function*() {
+      yield Cat.create({ _id: 'test' });
+      const res = yield Cat.findOneAndUpdate({}, {
+        $addToSet: { myArr: ['Case SenSiTive'] }
+      }, { new: true });
+      assert.equal(res.myArr[0], 'case sensitive');
     });
   });
 });